moved ILLEGAL_CHARS_REGEX
