+++ wiggle_query_language/clauses/regexes/make/tests/test_make_re.py
@@ -4,7 +4,6 @@ import pytest
 
 from testing.test_helpers import does_not_raise
 from wiggle_query_language.clauses.regexes.make.make_patterns import (
-    ILLEGAL_CHARS_REGEX,
     MAKE_STATEMENT_ALL_REGEX,
     MAKE_STATEMENT_CHECK_CLAUSE_SYNTAX_REGEX,
 )
@@ -61,28 +60,4 @@ class TestMakePatterns:
     ) -> None:
         with exception:
             test = MAKE_STATEMENT_CHECK_CLAUSE_SYNTAX_REGEX.findall(test_pattern)
-            assert test == expected_result
-
-    @pytest.mark.parametrize(
-        "test_pattern, expected_result, exception",
-        [
-            pytest.param(
-                """MAKE (:NodeLabel$)-[#:%]->(foo:NodeLabel*);""",
-                ["$", "#", "%", "*"],
-                does_not_raise(),
-                id="EXP PASS: 1 Match",
-            ),
-            pytest.param(
-                """MAKE (:NodeLabel)-[:]->(foo:NodeLabel);""",
-                [],
-                does_not_raise(),
-                id="EXP PASS: No Match",
-            ),
-        ],
-    )
-    def test_illegal_chars_regex(
-        self, test_pattern: str, expected_result: Optional[list[str]], exception
-    ) -> None:
-        with exception:
-            test = ILLEGAL_CHARS_REGEX.findall(test_pattern)
             assert test == expected_result
