+++ wiggle_query_language/clauses/regexes/patterns/utilities.py
@@ -1,0 +1,6 @@
+from re import IGNORECASE, compile
+
+from wiggle_query_language.clauses.regexes import ILLEGAL_CHARS
+
+# #%&*
+ILLEGAL_CHARS_REGEX = compile(pattern=rf"[{ILLEGAL_CHARS}]", flags=IGNORECASE)
